Bug fixes and licence updates
Changed over to the GNU LESSER GENERAL PUBLIC LICENSE version 2.1.

This has been done as this module will likely sit at the top level of any environment it is deployed within and I'm not a lawyer.

Fixed a bug where node did not garbage collect sockets fast enough which resulted in node freaking out. (Kinda reminded me of C)

Added the ability to redirect to domains outside of localhost again.

Edit the notice files and added the ability to view the licence of this project from the web interface.

Fixed broken styles on provided 404 page. Renamed server.ts to proxy.ts.

Updated readme